Move CSCConstants.h to DataFormats/CSCDigi
