+++ DataFormats/L1TMuon/src/EMTFHit.cc
@@ -1,5 +1,5 @@
 #include "DataFormats/L1TMuon/interface/EMTFHit.h"
-#include "DataFormats/L1TMuon/interface/CSCConstants.h"
+#include "DataFormats/CSCDigi/interface/CSCConstants.h"
 
 namespace l1t {
