+++ EventFilter/L1TRawToDigi/src/OmtfCscUnpacker.cc
@@ -1,11 +1,8 @@
 #include "EventFilter/L1TRawToDigi/interface/OmtfCscUnpacker.h"
-
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
-
 #include "DataFormats/MuonDetId/interface/CSCDetId.h"
 #include "DataFormats/CSCDigi/interface/CSCCorrelatedLCTDigi.h"
-#include "DataFormats/L1TMuon/interface/CSCConstants.h"
-
+#include "DataFormats/CSCDigi/interface/CSCConstants.h"
 #include "DataFormats/L1TMuon/interface/OMTF/OmtfCscDataWord64.h"
 
 namespace omtf {
