+++ L1Trigger/CSCTrackFinder/src/CSCSectorReceiverMiniLUT.cc
@@ -1,19 +1,19 @@
-#include <L1Trigger/CSCTriggerPrimitives/interface/CSCPatternBank.h>
-#include <DataFormats/L1CSCTrackFinder/interface/CSCBitWidths.h>
-#include <DataFormats/L1CSCTrackFinder/interface/CSCTFConstants.h>
-#include <DataFormats/L1TMuon/interface/CSCConstants.h>
+#include "L1Trigger/CSCTrackFinder/interface/CSCSectorReceiverMiniLUT.h"
+#include "L1Trigger/CSCTriggerPrimitives/interface/CSCPatternBank.h"
+#include "DataFormats/L1CSCTrackFinder/interface/CSCBitWidths.h"
+#include "DataFormats/L1CSCTrackFinder/interface/CSCTFConstants.h"
+#include "DataFormats/CSCDigi/interface/CSCConstants.h"
 
-#include <Geometry/CSCGeometry/interface/CSCLayerGeometry.h>
+#include "Geometry/CSCGeometry/interface/CSCLayerGeometry.h"
 #include "DataFormats/GeometryVector/interface/LocalPoint.h"
 #include "DataFormats/GeometryVector/interface/GlobalPoint.h"
 
-#include <DataFormats/MuonDetId/interface/CSCTriggerNumbering.h>
+#include "DataFormats/MuonDetId/interface/CSCTriggerNumbering.h"
 
-#include <FWCore/MessageLogger/interface/MessageLogger.h>
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
 
 #include <fstream>
 #include <cmath>
-#include <L1Trigger/CSCTrackFinder/interface/CSCSectorReceiverMiniLUT.h>
 
 lclphidat CSCSectorReceiverMiniLUT::calcLocalPhiMini(unsigned theadd, const bool gangedME1a) {
   // This method is ripped from CSCSectorReceverLUT.cc with minor changes
