+++ L1Trigger/CSCTriggerPrimitives/interface/LCTContainer.h
@@ -10,7 +10,7 @@
  */
 
 #include "DataFormats/CSCDigi/interface/CSCCorrelatedLCTDigi.h"
-#include "DataFormats/L1TMuon/interface/CSCConstants.h"
+#include "DataFormats/CSCDigi/interface/CSCConstants.h"
 
 #include <vector>
 #include <algorithm>
