+++ L1Trigger/CSCTriggerPrimitives/src/CSCMuonPortCard.cc
@@ -1,5 +1,5 @@
 #include "L1Trigger/CSCTriggerPrimitives/interface/CSCMuonPortCard.h"
-#include "DataFormats/L1TMuon/interface/CSCConstants.h"
+#include "DataFormats/CSCDigi/interface/CSCConstants.h"
 #include <algorithm>
 
 CSCMuonPortCard::CSCMuonPortCard() {}
